Change the ActionInvoker reference to the RenderedAdaptiveCard to a weakref.
Because the RenderedAdaptiveCard holds on to the root of the tree.
The ClickHandler for the action button hangs on to the ActionInvoker.
This causes a cycle cause the whole AdaptiveCard and XAML tree to leak.
By switching to a WeakRef on the ActionInvoker, this will allow the RenderedAdaptiveCard to be released and the whole tree to go away.

Also setting the RenderContext to use a weakRef on the RenderedAdaptiveCard to prevent another leak.